+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/ExtensionsSection/useExtensionsSection.tsx
@@ -1,6 +1,7 @@
 import { useContext } from 'react'
 import { TemplateSelectDialogContext } from '../context'
 import { useSelectionList } from '../useSelectionList'
+import { useRowRenderer } from './useRowRenderer'
 import { useColumns } from './useColumns'
 
 export function useExtensionsSection() {
@@ -14,5 +15,6 @@ export function useExtensionsSection() {
   )
   const items = context.props.data.extensions.filter((ext) => !ext.hidden)
   const columns = useColumns()
-  return { selection, selectedKeys, items, columns, onSearch, searchTerm } as const
+  const onRenderRow = useRowRenderer({ selectedKeys, searchTerm })
+  return { selection, items, columns, onSearch, onRenderRow } as const
 }
