+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/ListContentSection/index.tsx
@@ -1,15 +1,11 @@
 import {
-  DetailsList,
-  IDetailsRowProps,
-  ScrollablePane,
+  DetailsList, ScrollablePane,
   SelectionMode,
   Sticky,
   StickyPositionType
 } from '@fluentui/react'
-import { ListContentConfig } from 'models'
 import strings from 'ProjectExtensionsStrings'
 import React, { useContext } from 'react'
-import { CheckLocked } from '../CheckLocked'
 import { TemplateSelectDialogContext } from '../context'
 import { ListHeaderSearch } from '../ListHeaderSearch'
 import { TemplateListContentConfigMessage } from '../TemplateListContentConfigMessage'
@@ -19,7 +15,7 @@ import { useListContentSection } from './useListContentSection'
 
 export const ListContentSection: TemplateSelectDialogSectionComponent = (props) => {
   const context = useContext(TemplateSelectDialogContext)
-  const { selection, selectedKeys, items, columns, onSearch, searchTerm } = useListContentSection()
+  const { selection, items, columns, onSearch, onRenderRow } = useListContentSection()
 
   return (
     <div className={styles.root} style={props.style}>
@@ -29,33 +25,7 @@ export const ListContentSection: TemplateSelectDialogSectionComponent = (props) 
           selection={selection}
           selectionMode={SelectionMode.multiple}
           selectionPreservedOnEmptyClick={true}
-          onRenderRow={(
-            detailsRowProps: IDetailsRowProps,
-            defaultRender: (props?: IDetailsRowProps) => JSX.Element
-          ) => {
-            const lcc = detailsRowProps.item as ListContentConfig
-            const isLocked = lcc.isLocked(context.state.selectedTemplate)
-            detailsRowProps.disabled = isLocked
-            if (isLocked) {
-              detailsRowProps.onRenderCheck = (props) => (
-                <CheckLocked {...props} tooltip={{ text: strings.ListContentLockedTooltipText }} />
-              )
-              if (lcc.isDefault) {
-                detailsRowProps.styles = {
-                  root: { background: 'rgb(237, 235, 233)', color: 'rgb(50, 49, 48)' }
-                }
-              }
-            }
-            if (
-              lcc.text.toLowerCase().indexOf(searchTerm.toLowerCase()) !== -1 ||
-              selectedKeys.includes(lcc.key) ||
-              (isLocked && lcc.isDefault)
-            ) {
-              return defaultRender(detailsRowProps)
-            } else {
-              return null
-            }
-          }}
+          onRenderRow={onRenderRow}
           onRenderDetailsHeader={(detailsHeaderProps, defaultRender) => (
             <ListHeaderSearch
               detailsHeaderProps={detailsHeaderProps}
