+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/ListContentSection/useListContentSection.tsx
@@ -2,6 +2,7 @@ import { useContext } from 'react'
 import { TemplateSelectDialogContext } from '../context'
 import { useSelectionList } from '../useSelectionList'
 import { useColumns } from './useColumns'
+import { useRowRenderer } from './useRowRenderer'
 
 export function useListContentSection() {
   const context = useContext(TemplateSelectDialogContext)
@@ -14,5 +15,6 @@ export function useListContentSection() {
   )
   const items = context.props.data.listContentConfig.filter((lcc) => !lcc.hidden)
   const columns = useColumns()
-  return { selection, selectedKeys, items, columns, onSearch, searchTerm } as const
+  const onRenderRow = useRowRenderer({ selectedKeys, searchTerm })
+  return { selection, items, columns, onSearch, onRenderRow } as const
 }
